requirements.txt
Flask==3.0.2
pytz==2024.1
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
alembic==1.13.1
Flask-CORS==4.0.0
Flask-Talisman==1.1.0
Flask-Migrate==4.0.5
flask-swagger-ui==4.11.1
gunicorn==21.2.0
pytest==8.0.1
